Dont lose your mind

+++ ISA-Projekat/src/main/java/code/service/impl/CottageServiceImpl.java
@@ -167,11 +167,28 @@ public class CottageServiceImpl implements CottageService {
 
     @Override
     @Transactional
-    public void unlinkReferencesAndDeleteCottage(Integer id) throws EntityNotFoundException, EntityNotDeletableException {
-        checkIfCottageDeletable(id);
+    public void unlinkReferencesAndDeleteCottage(Integer id) throws EntityNotFoundException, EntityNotDeletableException, UnauthorizedAccessException, UserNotFoundException, EntityNotOwnedException {
+
+        Authentication auth;
+        try{
+            auth = SecurityContextHolder.getContext().getAuthentication();
+        }catch(Exception ex){
+            throw new UnauthorizedAccessException("Unauthorized");
+        }
+        CottageOwner owner;
+        try{
+            owner = (CottageOwner) auth.getPrincipal();
+        }
+        catch(ClassCastException ex){
+            throw new UnauthorizedAccessException("User is not a cottage owner");
+        }
+        if(owner == null) throw new UserNotFoundException("Cottage owner not found");
         Optional<Cottage> optionalCottage = _cottageRepository.findById(id);
         if(!optionalCottage.isPresent())throw new EntityNotFoundException("Cottage not found");
         Cottage cottage = optionalCottage.get();
+        if(cottage.getCottageOwner().getId() != owner.getId())throw new EntityNotOwnedException("Cottage not owned by given user");
+
+        checkIfCottageDeletable(id);
         for(Picture pic : cottage.getPictures()){
             FileUploadUtil.deleteFile(COTTAGE_PICTURE_DIRECTORY, cottage.getId() + "_" + pic.getName());
         }
